requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
datefinder==0.7.1
elasticsearch==7.10.1
idna==2.10
joblib==1.0.0
kafka-python==2.0.2
langdetect==1.0.8
nltk==3.5
oauthlib==3.1.0
py4j==0.10.7
PySocks==1.7.1
pyspark==2.4.7
python-dateutil==2.8.1
pytz==2020.5
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
textblob==0.15.3
tqdm==4.54.1
tweepy==3.9.0
urllib3==1.26.2
